fix: unify article parsing
